Altered File structure, hopefully added 64-bit file support on linux, changed playFile()
The file structure of repository was changed in order to allow a standard library to be easily added to this repository. Hopefully, large files (64-bit files) are now able to be loaded in Linux. The function, playFile() was changed to allow files relative to the main file, and relative to the interpreter's executable.